Switch disabling class to is-frozen
In the lesson player we've bound adding the class is-disabled and the attribute
disabled together. I think this is a good thing. But, there are times that we
want to disable dragging but still allow click events to fire on the non-
draggable element, so I'm changing this to is-frozen here.

It's a compromise. Alternate suggestions welcome.